+++ docs/CHANGELOG.md
@@ -1,3 +1,4 @@
+- #1104 Add `deposit` method to prover incentives module.
 - #1003 Removes the `BlobData` enum, requiring sequencers to post the `Batch` struct directly on chain.
 - #1062 Adds `remove` method to `StateVec`
 - #978 Upgrades rust toolchain to 1.79.